requirements.txt
django
djangorestframework
networkx
requests